optimization/requirements.txt
# Blood Bank Inventory Optimization Microservice
fastapi==0.104.1
uvicorn==0.24.0
pulp==2.7.0
scipy==1.11.4
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.0
requests==2.31.0
apscheduler==3.10.4
redis==5.0.1
motor==3.3.2
pymongo==4.6.0
aiohttp==3.8.6  # Changed from 3.9.1 to support Python 3.9
PyJWT==2.8.0
python-multipart==0.0.6
